requirements.txt
requests==2.32.3
pytest==8.3.5
flake8==7.1.1
google-cloud-storage==2.19.0